interval: float64->int64

+++ bow/window.go
@@ -4,8 +4,8 @@ type Window struct {
 	Bow                 Bow
 	FirstIndex          int // index (across all windows) of first row in this window (-1 if none)
 	IntervalColumnIndex int
-	Start               float64
-	End                 float64
+	Start               int64
+	End                 int64
 	IsInclusive         bool
 }
